Add a real implementation of `won`.
Woohoo, this is a real and fully featured tic-tac-toe game now! In a
real implementation, we'd probably want the board to be an array for
performance reasons, but doing it as all vecs makes things a bit
trickier, since we can't assume that the board is Copy (and similar to
the comment about Move in Magic: The Gathering, a board in MTG would not
be copy or compile time sized).